refactor(RandomLottoNumberGenerator -> LottoNumberGenerator)
