+++ README.md
@@ -57,4 +57,6 @@
 - [X] 로또 숫자 오름차순으로 정렬
 - [X] 랜덤 숫자 생성기
 - [X] 랜덤 숫자 생성기는 무작위로 로또 숫자들 중 6개를 뽑는다
+- [] 로또 생성기
+- [] 로또 생성기는 
 - [] 당첨금 계산기
